refactor column filter
