+++ app-test/collection/table-filter.php
@@ -42,7 +42,7 @@ $filter = $table->addFilter(new Table\Filter());
 $filter->addColumnFilter(new Table\Filter\Text('name'));
 $filter->addColumnFilter(new Table\Filter\Text('iso'));
 $filter->addColumnFilter(new Table\Filter\Text('iso3'));
-$filter->addColumnFilter(new Table\Filter\Number('numcode'));
+$filter->addColumnFilter(new Table\Filter\Integer('numcode'));
 
 // Response to an onDataRequest event from Table.
 // Fill in Table\Result\Set $dataSet depending on $payload value.
