+++ src/Component/Table/Filter/Generic.php
@@ -1,0 +1,50 @@
+<?php
+
+declare(strict_types = 1);
+/**
+ *
+ */
+
+namespace Fohn\Ui\Component\Table\Filter;
+
+use Fohn\Ui\Component\Table\Filter;
+
+abstract class Generic
+{
+    protected string $type = 'text';
+    private string $id;
+    protected string $label;
+
+    protected array $props = [
+        'type' => 'text',
+    ];
+
+    public function __construct(string $id, string $label = null, array $props = [])
+    {
+        $this->id = $id;
+        $this->label = $label ?? ucfirst($id);
+        $props['name'] = $this->id;
+        $this->props = array_merge($this->props, $props);
+    }
+
+    /** @return mixed */
+    abstract public function getValue(string $value);
+
+    public function getId(): string
+    {
+        return $this->id;
+    }
+
+    public function getDefinition(): array
+    {
+        return [
+            'id' => $this->id,
+            'label' => $this->label,
+            'operatorType' => $this->type,
+            'component' => [
+                'name' => Filter::getComponentName($this->type),
+                'props' => $this->props,
+            ],
+        ];
+    }
+}
